Silence compiler warning
git-svn-id: https://svn.modelica.org/projects/Modelica/trunk@8575 7ce873d0-865f-4ce7-a662-4bb36ea78beb

+++ Modelica/Resources/C-Sources/ModelicaStrings.c
@@ -468,7 +468,7 @@ MODELICA_EXPORT int ModelicaStrings_hashString(const char* str) {
 
     unsigned int hash = 0xAAAAAAAA;
     unsigned int i    = 0;
-    unsigned int len  = strlen(str);
+    unsigned int len  = (unsigned int)strlen(str);
 
     union hash_tag {
         unsigned int iu;
